Moved checkNodeValidity into common
