+++ src/renderer/components/node/NodeFooter/NodeFooter.tsx
@@ -1,8 +1,8 @@
 import { Center, SimpleGrid } from '@chakra-ui/react';
 import { memo } from 'react';
 import { useContextSelector } from 'use-context-selector';
+import { Validity } from '../../../../common/checkNodeValidity';
 import { GlobalVolatileContext } from '../../../contexts/GlobalNodeState';
-import { Validity } from '../../../helpers/checkNodeValidity';
 import { UseDisabled } from '../../../hooks/useDisabled';
 import { DisableToggle } from './DisableToggle';
 import { Timer } from './Timer';
